+++ packages/frontend/src/components/MkSignupDialog.form.vue
@@ -120,6 +120,7 @@ const props = withDefaults(defineProps<{
 		signupReasonRequired: true,
 		signupReasonMaxLength: 4096,
 		emojiRequestEnabled: false,
+		relayTimelineEnabled: false,
 	}),
 });