+++ packages/frontend/src/components/MkStreamingNotesTimeline.vue
@@ -80,7 +80,7 @@ import { isSeparatorNeeded, getSeparatorInfo } from '@/utility/timeline-date-sep
 import { Paginator } from '@/utility/paginator.js';
 
 const props = withDefaults(defineProps<{
-	src: BasicTimelineType | 'mentions' | 'directs' | 'list' | 'antenna' | 'channel' | 'role';
+	src: BasicTimelineType | 'mentions' | 'directs' | 'list' | 'antenna' | 'channel' | 'role' | 'relay';
 	list?: string;
 	antenna?: string;
 	channel?: string;
@@ -178,6 +178,14 @@ if (props.src === 'antenna') {
 	paginator = markRaw(new Paginator('roles/notes', {
 		computedParams: computed(() => ({
 			roleId: props.role!,
+		})),
+		useShallowRef: true,
+	}));
+} else if (props.src === 'relay') {
+	paginator = markRaw(new Paginator('notes/relay-timeline', {
+		computedParams: computed(() => ({
+			withRenotes: props.withRenotes,
+			withFiles: props.onlyFiles ? true : undefined,
 		})),
 		useShallowRef: true,
 	}));
@@ -316,6 +324,7 @@ const connections = {
 	userList: null as Misskey.IChannelConnection<Misskey.Channels['userList']> | null,
 	channel: null as Misskey.IChannelConnection<Misskey.Channels['channel']> | null,
 	roleTimeline: null as Misskey.IChannelConnection<Misskey.Channels['roleTimeline']> | null,
+	relayTimeline: null as Misskey.IChannelConnection<Misskey.Channels['relayTimeline']> | null,
 };
 
 function connectChannel() {
@@ -383,6 +392,12 @@ function connectChannel() {
 			roleId: props.role,
 		});
 		connections.roleTimeline.on('note', prepend);
+	} else if (props.src === 'relay') {
+		connections.relayTimeline = stream.useChannel('relayTimeline', {
+			withRenotes: props.withRenotes,
+			withFiles: props.onlyFiles ? true : undefined,
+		});
+		connections.relayTimeline.on('note', prepend);
 	}
 }