+++ packages/misskey-js/src/streaming.types.ts
@@ -109,6 +109,17 @@ export type Channels = {
 		};
 		receives: null;
 	};
+	/** JUICE */
+	relayTimeline: {
+		params: {
+			withRenotes?: boolean;
+			withFiles?: boolean;
+		};
+		events: {
+			note: (payload: Note) => void;
+		};
+		receives: null;
+	};
 	userList: {
 		params: {
 			listId: string;